# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Landing Page Cursos - API",
        "description": "API REST desenvolvida em Laravel 11 para gerenciar cursos e competências",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API"
        }
    ],
    "paths": {
        "/courses": {
            "get": {
                "tags": ["Cursos"],
                "summary": "Listar cursos ativos",
                "description": "Retorna todos os cursos ativos com seus módulos, competências e informações de decisão",
                "operationId": "75403048c36d79fe464ce4716c204c61",
                "responses": {
                    "200": {
                        "description": "Lista de cursos ativos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Cursos"],
                "summary": "Criar novo curso",
                "description": "Cria um novo curso com os dados fornecidos",
                "operationId": "56ff5a96b473881afe594ca81aeb754b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": ["name", "description", "owner"],
                                "properties": {
                                    "name": {
                                        "description": "Nome do curso",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "description": "Descrição do curso",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "Tipo do curso",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "main_category": {
                                        "description": "Categoria principal",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "duration_hours": {
                                        "description": "Duração em horas",
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "modules_count": {
                                        "description": "Número de módulos",
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "access_period_months": {
                                        "description": "Período de acesso em meses",
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "modality": {
                                        "description": "Modalidade",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "price": {
                                        "description": "Preço",
                                        "type": "number",
                                        "format": "decimal",
                                        "minimum": 0
                                    },
                                    "enrollment_fee": {
                                        "description": "Taxa de matrícula",
                                        "type": "number",
                                        "format": "decimal",
                                        "minimum": 0
                                    },
                                    "max_installments": {
                                        "description": "Máximo de parcelas",
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "installment_value": {
                                        "description": "Valor da parcela",
                                        "type": "number",
                                        "format": "decimal",
                                        "minimum": 0
                                    },
                                    "featured": {
                                        "description": "Curso em destaque",
                                        "type": "boolean"
                                    },
                                    "active": {
                                        "description": "Curso ativo",
                                        "type": "boolean"
                                    },
                                    "target_audience": {
                                        "description": "Público-alvo",
                                        "type": "string"
                                    },
                                    "payment_info": {
                                        "description": "Informações de pagamento",
                                        "type": "string"
                                    },
                                    "payment_conditions": {
                                        "description": "Condições de pagamento",
                                        "type": "string"
                                    },
                                    "instructor": {
                                        "description": "Instrutor",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "level": {
                                        "description": "Nível",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "category": {
                                        "description": "Categoria",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "owner": {
                                        "description": "Proprietário",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "image": {
                                        "description": "Imagem do curso",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "banner": {
                                        "description": "Banner do curso",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Curso criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/courses/{course}": {
            "get": {
                "tags": ["Cursos"],
                "summary": "Exibir curso específico",
                "description": "Retorna os detalhes de um curso específico com seus módulos, competências e informações de decisão",
                "operationId": "10f1a3f466f9f189b72f4faa087511d1",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID do curso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do curso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Curso não encontrado"
                    }
                }
            },
            "delete": {
                "tags": ["Cursos"],
                "summary": "Excluir curso",
                "description": "Remove um curso do sistema",
                "operationId": "b3c52c7b9051fe701b0428f229852379",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID do curso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Curso excluído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Curso não encontrado"
                    }
                }
            }
        },
...
    },
...
}